NXIO-375: Secure etcd cleaner

+++ fleet/cleaner/etcd-cleaner.sh
@@ -1,6 +1,11 @@
 #!/bin/sh -
 
-DOMAIN=`etcdctl get /services/${SERVICE_ID}/1/domain`
-
+if [ -z "${SERVICE_ID}" ]; then
+  exit 1
+fi
 etcdctl rm --recursive /services/${SERVICE_ID}
-etcdctl rm --recursive /domains/${DOMAIN}
+
+DOMAIN=`etcdctl get /services/${SERVICE_ID}/1/domain`
+if [ $? -eq 0 ]; then
+  etcdctl rm --recursive /domains/${DOMAIN}
+fi
